+++ test/unit/test_structured_grids.py
@@ -74,7 +74,7 @@ class TestCartGridGeometry2DPert(unittest.TestCase):
 class TestCartGridGeometryUnpert3D(unittest.TestCase):
 
     def setUp(self):
-        nc = 2 * np.ones(3)
+        nc = 2 * np.ones(3, dtype=np.int)
         g = structured.CartGrid(nc)
         g.compute_geometry()
         self.g = g
@@ -135,7 +135,7 @@ class TestCartGridGeometryUnpert3D(unittest.TestCase):
 class TestCartGridGeometry1CellPert3D(unittest.TestCase):
 
     def setUp(self):
-        nc = np.ones(3)
+        nc = np.ones(3, dtype=np.int)
         g = structured.CartGrid(nc)
         g.nodes[:, -1] = [2, 2, 2]
         g.compute_geometry()
